requirements/requirements.txt
Django>=3.2.13,<3.13
django-choices>=1.7.2,<1.8
django-cors-headers>=3.12.0,<3.13
django-environ>=0.9.0,<0.10
django-extensions>=3.1.5,<3.2
django-filter>=2.4.0,<2.5
django-property-filter>=1.1.2,<1.2
django-health-check>=3.16.5,<3.17
django-redis>=5.2.0,<5.3
django-split-settings>=1.1.0,<1.2
django-utils-six>=2.0,<2.1
djangorestframework>=3.13.1,<3.14
djangorestframework-api-key>=2.2.0,<2.3
djangorestframework-simplejwt>=5.2.0,<5.3
drf-access-policy>=1.1.1,<1.2
drf_spectacular>=0.25.1,<0.26
drf_nested_routers>=0.93.3,<0.94
gunicorn>=20.1.0,<20.2
channels>=3.0.4,<3.1
channels-redis>=3.4.0,<3.5
psycopg2-binary>=2.9.3,<2.10
uvicorn>=0.17.6,<0.18
pyotp>=2.8.0,<2.9
qrcode>=7.3.1,<7.4
factory_boy>=3.2.1,<3.3
black>=23.1.0,<23.2